requirements.txt
python-dotenv~=1.0.1
uuid~=1.30
# pymupdf~=1.24.5
Pillow~=10.3.0
# git+https://github.com/huggingface/transformers
# huggingface_hub~=0.23.3
# chardet~=5.2.0
# accelerate~=0.31.0
# bitsandbytes~=0.43.1
# sentence_transformers~=3.0.1
# chromadb~=0.5.0
# argparse~=1.4.0
# tqdm~=4.66.4
pyautogui~=0.9.54
# openai~=1.34.0
# pydantic~=2.7.4
# ollama~=0.2.1
streamlit~=1.35.0
opencv-python~=4.10.0.82
numpy<2